Changed move integers to bytes. (#111)

+++ src/PokeGame.Core/Moves/Power.cs
@@ -4,15 +4,15 @@ namespace PokeGame.Core.Moves;
 
 public sealed class Power
 {
-  public int Value { get; }
+  public byte Value { get; }
 
-  public Power(int value)
+  public Power(byte value)
   {
     Value = value;
     new Validator().ValidateAndThrow(this);
   }
 
-  public static Power? TryCreate(int? value) => value.HasValue ? new(value.Value) : null;
+  public static Power? TryCreate(byte? value) => value.HasValue ? new(value.Value) : null;
 
   public override bool Equals(object? obj) => obj is Power power && power.Value == Value;
   public override int GetHashCode() => Value.GetHashCode();
