+++ src/PokeGame.Infrastructure/Converters/AccuracyConverter.cs
@@ -6,7 +6,7 @@ internal class AccuracyConverter : JsonConverter<Accuracy>
 {
   public override Accuracy? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
   {
-    return reader.TryGetInt32(out int value) ? new Accuracy(value) : null;
+    return reader.TryGetByte(out byte value) ? new Accuracy(value) : null;
   }
 
   public override void Write(Utf8JsonWriter writer, Accuracy accuracy, JsonSerializerOptions options)
